Add files via upload
As the world's largest retailer, Walmart relies heavily on data to make informed business decisions. That's why I'm excited to share about the new Walmart sales dashboard that has recently been developed.

This sales dashboard is a powerful tool that provides Walmart executives and store managers with real-time insights into sales performance across the entire company. By integrating data from Walmart's point-of-sale systems and other sources, the dashboard can quickly and easily identify trends and patterns in sales data, allowing decision-makers to adjust their strategies accordingly.

One of the key benefits of the Walmart sales dashboard is its ability to provide detailed insights into customer behavior. With this tool, Walmart can analyze which products are selling the most, which stores are performing the best, and even which regions are seeing the most sales growth. This information can then be used to make data-driven decisions about inventory management, pricing, and marketing.

In addition to its analytical capabilities, the Walmart sales dashboard is also incredibly user-friendly. With intuitive charts, graphs, and other visualizations, users can quickly and easily understand the data and make informed decisions. And because the dashboard is cloud-based, it can be accessed from anywhere, allowing executives and managers to stay up-to-date on sales performance even when they're out of the office.

Overall, the Walmart sales dashboard is a powerful tool that is sure to have a significant impact on Walmart's sales strategy moving forward. By providing real-time insights into sales performance and customer behavior, this tool is helping Walmart stay ahead of the curve in an increasingly competitive retail landscape.


The data includes 12 columns and 3177 rows from 10 different states from USA from year 2011-2015 with detail information about the products from 17 different categories.